Separate out reusable components for the sync modals
