+++ src/page-multi-facility/FacilityInputForm.tsx
@@ -27,7 +27,7 @@ import AddCasesModal from "./AddCasesModal";
 import FacilityProjections from "./FacilityProjections";
 import HistoricalCasesChart from "./HistoricalCasesChart";
 import LocaleInformationSection from "./LocaleInformationSection";
-import SyncReferenceFacilityModal from "./SyncReferenceFacilityModal";
+import SyncNewFacility from "./ReferenceDataModal/SyncNewFacility";
 import { Facility, Scenario } from "./types";
 
 interface ButtonSectionProps {
@@ -321,7 +321,7 @@ const FacilityInputForm: React.FC<Props> = ({ scenarioId }) => {
         </Column>
 
         {/* MODAL */}
-        <SyncReferenceFacilityModal
+        <SyncNewFacility
           facilityId={syncDataModalFacility}
           onClose={navigateHome}
         />
